Remove the list of decoration classes from the XBL/HTC decorator URIs. The list is now figured out by introspecting the element (which may be identified for decoration using any number of means).
The decorator maintains a set of rules that it applies to each element that needs to be decorated. Rules have the following properties:
 * name - Used to identify the rule (not currently used)
 * match - The optional function that returns true iff the candidate element needs to be decorated by the rule (paragraph below explains why it is optional)
 * apply - The function that manipulates the list of class names

For performance, the rules are first partitioned by namespace, and further by element local name. The match() function is optional since if the only criteria for matching is element QName, that is handled by the partitioning. A missing match() is the same as always returning true.

For rules that may be applicable across element names within a namespace (for example, the presence of a "ref" attribute in XForms mark-up), a "wildcard" partition is used (we have no such rules at present). There is also a wildcard partition to match any element across any namespace (expected to be used sparingly, if at all).

A public method DECORATOR.addDecorationRules() has been added:
 * The rules for a given namespace can be added in a "distributed" manner i.e. multiple DECORATOR.addDecorationRules() calls can be made for a given namespace. The result is to concatenate the rules for each element local name (and wildcard) added across calls.
 * The XForms specific rules are added in xforms-defs.js

Related minor changes to decorate.js:
(1) The generateMozBindingStyle method will only generate the ? character if the 'objects' array is not empty
(2) The getDecorationObjectNames method will never split an empty string (this results in a one element array instead of an empty array)

Related minor change to namespaceManager.js:
(1) A NamespaceManager.getNamespaceURI(node) function has been added (does what the name suggests).