Added support for the XForms hint element.
This was added in two stages.

The first is a new module called document.notify, which will contain all notification code. At the moment it just contains ephemeral message support, but eventually it will provide support for debug logging, modal and modeless dialogs (via the Ubiquity Message library), and anything else we can think of.

The second stage was to then use the ephemeral message function in the the hint event handlers, attached to any control that contains a hint element.

Note that the default format for hints is in 'default.css', but the CSS rules that are used to actually implement hint (such as positioning and setting the parent element) are in a new file called 'lib/xforms/style/hints.css'.